programs/article_summarizer_backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sse-starlette==1.8.2

# HTTP and scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Playwright for browser automation
playwright>=1.40.0

# Video/audio processing
youtube-transcript-api>=0.6.0
yt-dlp>=2023.0.0
pydub>=0.25.0
imagehash>=4.3.0
pillow>=10.0.0

# Database
supabase>=2.0.0

# AI APIs
anthropic>=0.8.0
openai>=1.0.0
deepgram-sdk>=5.2.0
braintrust>=0.3.5

# Utilities
python-dotenv>=1.0.0
